backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.2.0
psycopg2-binary==2.9.9
GDAL==3.4.3
django-environ==0.11.2
python-decouple==3.8
Pillow==10.2.0
redis==5.0.1
africastalking==1.2.5
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
